# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ai/{providerId}/models:
        get:
            tags:
                - AIService
            description: |-
                ListProviderModels lists the models an instance AI provider offers. The
                 server calls the provider's model catalog endpoint, so this requires
                 network access to the provider.
            operationId: AIService_ListProviderModels
            parameters:
                - name: providerId
                  in: path
                  description: Required. The provider id from the instance AI setting.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListProviderModelsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:analyzeMemoSemantic:
        post:
            tags:
                - AIService
            description: AnalyzeMemoSemantic returns ephemeral Jev judgments for one readable memo.
            operationId: AIService_AnalyzeMemoSemantic
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AnalyzeMemoSemanticRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AnalyzeMemoSemanticResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:chat:
        post:
            tags:
                - AIService
            description: |-
                Chat runs one conversational turn with an instance AI provider. The caller
                 resends the whole conversation each turn; the server keeps no chat state.
                 The model may propose note changes, but this method never writes a memo.
            operationId: AIService_Chat
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ChatRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ChatResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:estimateChatContext:
        post:
            tags:
                - AIService
            description: |-
                EstimateChatContext resolves a context selection and reports how much of
                 the chat context budget it would consume. The Hub calls this while the user
                 edits the selection so an over-budget selection is refused before sending.
            operationId: AIService_EstimateChatContext
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EstimateChatContextRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EstimateChatContextResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:transcribe:
        post:
            tags:
                - AIService
            description: Transcribe transcribes an audio file using an instance AI provider.
            operationId: AIService_Transcribe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TranscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TranscribeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        AnalyzeMemoSemanticRequest:
            required:
                - memo
            type: object
            properties:
                memo:
                    type: string
                    description: Required. Memo resource name in the form memos/{uid}.
        AnalyzeMemoSemanticResponse:
            type: object
            properties:
                ideaProbability:
                    type: number
                    format: double
                taskProbability:
                    type: number
                    format: double
                journalProbability:
                    type: number
                    format: double
                referenceProbability:
                    type: number
                    format: double
                actionableProbability:
                    type: number
                    format: double
                revisitProbability:
                    type: number
                    format: double
                technicalProbability:
                    type: number
                    format: double
                questionProbability:
                    type: number
                    format: double
                timeSensitiveProbability:
                    type: number
                    format: double
                model:
                    type: string
                    description: Provider-reported resolved model.
...
        ChatRequest:
            required:
                - messages
            type: object
            properties:
                filter:
                    type: string
                    description: |-
                        Optional. A CEL filter selecting which notes the model may read. An empty
                         filter means the model sees none of the user's notes.
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatMessage'
                    description: |-
                        Required. The conversation so far, oldest first, ending with the new user
                         message. The server keeps no conversation state.
                includeComments:
                    type: boolean
                    description: |-
                        Optional. Whether each selected note's comment thread is injected alongside
                         it. Defaults to false. Comments enlarge the context, so a turn that enables
                         this can exceed a budget the same selection met without it.
        ChatResponse:
            type: object
            properties:
                content:
                    type: string
                    description: The model's reply, with proposal blocks removed.
                proposals:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatProposal'
                    description: The note changes the model proposed. Never applied by the server.
                contextMemoCount:
                    type: string
                    description: How many notes were injected as context.
                contextEstimatedTokens:
                    type: string
                    description: The estimated token cost of the injected context.
                contextBudgetTokens:
                    type: string
                    description: The context budget this turn was measured against.
                promptTokens:
                    type: string
                    description: Token usage reported by the provider. Zero when the provider omits it.
                completionTokens:
                    type: string
                totalTokens:
                    type: string
                model:
                    type: string
                    description: |-
                        The model the provider reports having used, which can differ from the
                         requested model when a router picks a backend.
                truncated:
                    type: boolean
                    description: True when the provider stopped because the reply hit the token limit.
                contextCommentCount:
                    type: string
                    description: |-
                        How many comments were injected alongside the notes. Zero unless the turn
                         enabled include_comments.
...
        EstimateChatContextRequest:
            required:
                - filter
            type: object
            properties:
                filter:
                    type: string
                    description: |-
                        Required. A CEL filter selecting which notes the model may read. An empty
                         filter selects nothing, not everything.
                includeComments:
                    type: boolean
                    description: |-
                        Optional. Whether each matched note's comment thread is injected alongside
                         it. Defaults to false, so a selection costs exactly what the estimate
                         reports unless the caller opts in. Comments are never selected directly:
                         they follow the notes the filter matched, under the caller's own access
                         scope, so a comment the caller cannot read is never included.
        EstimateChatContextResponse:
            type: object
            properties:
                memoCount:
                    type: string
                    description: How many notes the selection matched.
                totalChars:
                    type: string
                    description: The total character count of the matched notes' content.
                estimatedTokens:
                    type: string
                    description: |-
                        The estimated token cost of the selection. Comment content is not counted
                         here, because comments are resolved only when a turn is actually sent.
                contextBudgetTokens:
                    type: string
                    description: The budget this estimate was measured against.
                fits:
                    type: boolean
                    description: Whether the selection fits the budget.
...
        ListProviderModelsResponse:
            type: object
            properties:
                models:
                    type: array
                    items:
                        $ref: '#/components/schemas/AIProviderModel'
                    description: The models the provider offers, ordered by identifier.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TranscribeRequest:
            required:
                - audio
            type: object
            properties:
                audio:
                    allOf:
                        - $ref: '#/components/schemas/TranscriptionAudio'
                    description: Required. Audio input.
...